Create a separate MutByteArray module
